Support EntitySchemaId in EntityIdValue
Various places that previously expected a StringValue to represent an
EntitySchema ID can now also take an EntityIdValue, with the common code
extracted into a utility class. The “extract” method is static for now,
mainly for convenience; if necessary (e.g. to improve error handling,
maybe?), we can make this a proper service later.

The extractor returns an EntitySchemaId instance, even if all the
callers get the serialization again afterwards, because I think it’s
cleaner this way, it conveys more information in the return type, and
also this lets the EntitySchemaId constructor catch malformed IDs in a
StringValue.

TODO: EntitySchemaFormatterTest gained quite a few tests here back when
I still special-cased this in EntitySchemaFormatter – probably add some
tests for the other affected classes too? And the extractor itself ofc.

Change-Id: I41e39ba4562f7c7dd1d508d8bfbbcfbecf48cd0c